+++ app/digital/ts/models/ioobjects/other/Decoder.ts
@@ -9,9 +9,11 @@ import {ConstantSpacePositioner} from "core/models/ports/positioners/ConstantSpa
 import {DigitalComponent} from "digital/models/DigitalComponent";
 import {InputPort} from "digital/models/ports/InputPort";
 import {OutputPort} from "digital/models/ports/OutputPort";
+import {PortLabeler} from "digital/actions/ports/PortLabeler";
 
 @serializable("Decoder")
 export class Decoder extends DigitalComponent {
+    protected portLabeler: PortLabeler;
 
     public constructor() {
         super(new ClampedValue(2,1,8),
@@ -22,7 +24,8 @@ export class Decoder extends DigitalComponent {
 
         // activate 0th port for initial state
         super.activate(true, 0);
-        this.setBinaryLabels();
+        this.portLabeler = new PortLabeler(this);
+        this.portLabeler.setBinaryLabels();
     }
 
     public activate(): void {
@@ -43,21 +46,10 @@ export class Decoder extends DigitalComponent {
         const width = Math.max(2 * padding + val * 8.342285, DEFAULT_SIZE);
         this.transform.setSize(V(width, DEFAULT_SIZE/2*Math.pow(2, val)));
         super.setInputPortCount(val);
-        this.setBinaryLabels();
+        this.portLabeler.setBinaryLabels();
     }
 
     public getDisplayName(): string {
         return "Decoder";
-    }
-
-    public setBinaryLabels(): void {
-        const ports = this.getOutputPorts();
-        const digitCount = Math.log2(ports.length);
-        let numStr = "0".repeat(digitCount);
-
-        for (let i = 0; i < Math.pow(2, digitCount); i++) {
-            ports[i].setName(numStr);
-            numStr = (numStr.substr(0, numStr.lastIndexOf("0")) + "1").padEnd(digitCount, "0");
-        }
     }
 }
