+++ app/digital/ts/models/ioobjects/other/Encoder.ts
@@ -9,9 +9,11 @@ import {ConstantSpacePositioner} from "core/models/ports/positioners/ConstantSpa
 import {DigitalComponent} from "digital/models/DigitalComponent";
 import {InputPort} from "digital/models/ports/InputPort";
 import {OutputPort} from "digital/models/ports/OutputPort";
+import {PortLabeler} from "digital/actions/ports/PortLabeler";
 
 @serializable("Encoder")
 export class Encoder extends DigitalComponent {
+    protected portLabeler: PortLabeler;
 
     public constructor() {
         super(new ClampedValue(4,2,Math.pow(2,8)),
@@ -19,7 +21,9 @@ export class Encoder extends DigitalComponent {
               V(DEFAULT_SIZE, DEFAULT_SIZE*2),
               new ConstantSpacePositioner<InputPort>(DEFAULT_SIZE/2),
               new ConstantSpacePositioner<OutputPort>(DEFAULT_SIZE/2));
-        this.setBinaryLabels();
+        
+        this.portLabeler = new PortLabeler(this);
+        this.portLabeler.setBinaryLabels();
     }
 
     public activate(): void {
@@ -44,21 +48,11 @@ export class Encoder extends DigitalComponent {
         const width = Math.max(2 * padding + val * 8.342285, DEFAULT_SIZE);
         this.transform.setSize(V(width, DEFAULT_SIZE/2*Math.pow(2, val)));
         super.setOutputPortCount(val);
-        this.setBinaryLabels();
+        this.portLabeler.setBinaryLabels();
     }
 
     public getDisplayName(): string {
         return "Encoder";
     }
-
-    public setBinaryLabels(): void {
-        const ports = this.getInputPorts();
-        const digitCount = Math.log2(ports.length);
-        let numStr = "0".repeat(digitCount);
-
-        for (let i = 0; i < Math.pow(2, digitCount); i++) {
-            ports[i].setName(numStr);
-            numStr = (numStr.substr(0, numStr.lastIndexOf("0")) + "1").padEnd(digitCount, "0");
-        }
-    }
+    
 }
